More GLSL optims

+++ neo/renderer/glsl/diffuseCubeShaderVP.cpp
@@ -29,8 +29,8 @@ attribute vec3 attr_TexCoord;
 // Uniforms
 uniform highp mat4 u_modelViewProjectionMatrix;
 uniform mat4 u_textureMatrix;
-uniform float u_colorAdd;
-uniform float u_colorModulate;
+uniform lowp float u_colorAdd;
+uniform lowp float u_colorModulate;
   
 // Out
 // gl_Position
